# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AddMcpServerRequest": {
        "additionalProperties": false,
        "properties": {
          "qualifiedName": {
            "minLength": 1,
            "title": "Qualifiedname",
            "type": "string"
          }
        },
        "required": [
          "qualifiedName"
        ],
        "title": "AddMcpServerRequest",
        "type": "object"
      },
      "AdoptMcpRequest": {
        "additionalProperties": false,
        "properties": {
          "harnesses": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harnesses"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "sourceHarness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceharness"
          }
        },
        "required": [
          "name"
        ],
        "title": "AdoptMcpRequest",
        "type": "object"
      },
...
      "DisableMcpServerRequest": {
        "properties": {
          "harness": {
            "description": "Harness identifier",
            "minLength": 1,
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness"
        ],
        "title": "DisableMcpServerRequest",
        "type": "object"
      },
...
      "EnableMcpServerRequest": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "harness": {
            "description": "Harness identifier",
            "minLength": 1,
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness"
        ],
        "title": "EnableMcpServerRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "McpApplyConfigResponse": {
        "properties": {
          "failed": {
            "items": {
              "$ref": "#/components/schemas/McpMutationFailureResponse"
            },
            "title": "Failed",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "server": {
            "$ref": "#/components/schemas/McpServerSpecResponse"
          },
          "succeeded": {
            "items": {
              "type": "string"
            },
            "title": "Succeeded",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "server",
          "succeeded",
          "failed"
        ],
        "title": "McpApplyConfigResponse",
        "type": "object"
      },
      "McpAvailabilityCheckResponse": {
        "properties": {
          "availabilityReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availabilityreason"
          },
          "availabilityStatus": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Availabilitystatus",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "name",
          "availabilityStatus"
        ],
        "title": "McpAvailabilityCheckResponse",
        "type": "object"
      },
...
      "McpConfigChoiceResponse": {
        "properties": {
          "configPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configpath"
          },
          "env": {
            "items": {
              "$ref": "#/components/schemas/McpEnvEntryResponse"
            },
            "title": "Env",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "logoKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logokey"
          },
          "payloadPreview": {
            "additionalProperties": true,
            "title": "Payloadpreview",
            "type": "object"
          },
          "sourceHarness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceharness"
          },
          "sourceKind": {
            "enum": [
              "managed",
              "harness"
            ],
            "title": "Sourcekind",
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/McpServerSpecResponse"
          }
        },
        "required": [
          "sourceKind",
          "label",
          "payloadPreview",
          "spec"
        ],
        "title": "McpConfigChoiceResponse",
        "type": "object"
      },
      "McpEnvEntryResponse": {
        "properties": {
          "isEnvRef": {
            "title": "Isenvref",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "key",
          "isEnvRef"
        ],
        "title": "McpEnvEntryResponse",
        "type": "object"
      },
...
      "McpInventoryColumnResponse": {
        "properties": {
          "configPresent": {
            "title": "Configpresent",
            "type": "boolean"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "installed": {
            "title": "Installed",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "logoKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logokey"
          },
          "mcpUnavailableReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mcpunavailablereason"
          },
          "mcpWritable": {
            "default": true,
            "title": "Mcpwritable",
            "type": "boolean"
          }
        },
        "required": [
          "harness",
          "label",
          "installed",
          "configPresent"
        ],
        "title": "McpInventoryColumnResponse",
        "type": "object"
      },
      "McpInventoryEntryResponse": {
        "properties": {
          "availabilityReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availabilityreason"
          },
          "availabilityStatus": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Availabilitystatus",
            "type": "string"
          },
          "canEnable": {
            "title": "Canenable",
            "type": "boolean"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "enabledStatus": {
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Enabledstatus",
            "type": "string"
          },
          "kind": {
            "enum": [
              "managed",
              "unmanaged"
            ],
            "title": "Kind",
            "type": "string"
          },
          "mcpStatus": {
            "$ref": "#/components/schemas/McpStatusResponse"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sightings": {
            "items": {
              "$ref": "#/components/schemas/McpBindingResponse"
            },
            "title": "Sightings",
            "type": "array"
          },
          "spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/McpServerSpecResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "displayName",
          "kind",
          "canEnable",
          "enabledStatus",
          "availabilityStatus",
          "mcpStatus",
          "sightings"
        ],
        "title": "McpInventoryEntryResponse",
        "type": "object"
      },
      "McpInventoryIssueResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "name",
          "reason"
        ],
        "title": "McpInventoryIssueResponse",
        "type": "object"
      },
      "McpInventoryResponse": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/McpInventoryColumnResponse"
            },
            "title": "Columns",
            "type": "array"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/McpInventoryEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/McpInventoryIssueResponse"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "columns",
          "entries"
        ],
        "title": "McpInventoryResponse",
        "type": "object"
      },
...
      "McpMutationFailureResponse": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness",
          "error"
        ],
        "title": "McpMutationFailureResponse",
        "type": "object"
      },
      "McpServerDetailResponse": {
        "properties": {
          "availabilityReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availabilityreason"
          },
          "availabilityStatus": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Availabilitystatus",
            "type": "string"
          },
          "canEnable": {
            "title": "Canenable",
            "type": "boolean"
          },
          "configChoices": {
            "items": {
              "$ref": "#/components/schemas/McpConfigChoiceResponse"
            },
            "title": "Configchoices",
            "type": "array"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "enabledStatus": {
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Enabledstatus",
            "type": "string"
          },
          "env": {
            "items": {
              "$ref": "#/components/schemas/McpEnvEntryResponse"
            },
            "title": "Env",
            "type": "array"
          },
          "kind": {
            "enum": [
              "managed",
              "unmanaged"
            ],
            "title": "Kind",
            "type": "string"
          },
          "marketplaceLink": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/McpMarketplaceLinkResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "mcpStatus": {
            "$ref": "#/components/schemas/McpStatusResponse"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sightings": {
            "items": {
              "$ref": "#/components/schemas/McpBindingResponse"
            },
            "title": "Sightings",
            "type": "array"
          },
          "spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/McpServerSpecResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "displayName",
          "kind",
          "canEnable",
          "enabledStatus",
          "availabilityStatus",
          "mcpStatus",
          "sightings"
        ],
        "title": "McpServerDetailResponse",
        "type": "object"
      },
      "McpServerMutationResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "server": {
            "$ref": "#/components/schemas/McpServerSpecResponse"
          }
        },
        "required": [
          "ok",
          "server"
        ],
        "title": "McpServerMutationResponse",
        "type": "object"
      },
      "McpServerSpecResponse": {
        "properties": {
          "args": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Args"
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "env": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Env"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "installedAt": {
            "title": "Installedat",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/McpSourceResponse"
          },
          "transport": {
            "enum": [
              "stdio",
              "http",
              "sse"
            ],
            "title": "Transport",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "name",
          "displayName",
          "source",
          "transport",
          "installedAt",
          "revision"
        ],
        "title": "McpServerSpecResponse",
        "type": "object"
      },
      "McpSetHarnessesResultResponse": {
        "properties": {
          "failed": {
            "items": {
              "$ref": "#/components/schemas/McpMutationFailureResponse"
            },
            "title": "Failed",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "succeeded": {
            "items": {
              "type": "string"
            },
            "title": "Succeeded",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "succeeded",
          "failed"
        ],
        "title": "McpSetHarnessesResultResponse",
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
      "ReconcileMcpServerRequest": {
        "additionalProperties": false,
        "properties": {
          "harnesses": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harnesses"
          },
          "sourceHarness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceharness"
          },
          "sourceKind": {
            "enum": [
              "managed",
              "harness"
            ],
            "title": "Sourcekind",
            "type": "string"
          }
        },
        "required": [
          "sourceKind"
        ],
        "title": "ReconcileMcpServerRequest",
        "type": "object"
      },
...
      "SetMcpServerHarnessesRequest": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "target": {
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target"
        ],
        "title": "SetMcpServerHarnessesRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "skill-manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/mcp/servers": {
      "get": {
        "operationId": "list_mcp_servers_api_mcp_servers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpInventoryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Mcp Servers"
      },
      "post": {
        "operationId": "install_mcp_server_api_mcp_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMcpServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install Mcp Server"
      }
    },
    "/api/mcp/servers/{name}": {
      "delete": {
        "operationId": "uninstall_mcp_server_api_mcp_servers__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpSetHarnessesResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Uninstall Mcp Server"
      },
      "get": {
        "operationId": "get_mcp_server_api_mcp_servers__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mcp Server"
      }
    },
    "/api/mcp/servers/{name}/availability/check": {
      "post": {
        "operationId": "check_mcp_server_availability_api_mcp_servers__name__availability_check_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpAvailabilityCheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Mcp Server Availability"
      }
    },
    "/api/mcp/servers/{name}/disable": {
      "post": {
        "operationId": "disable_mcp_server_api_mcp_servers__name__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableMcpServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Mcp Server"
      }
    },
    "/api/mcp/servers/{name}/enable": {
      "post": {
        "operationId": "enable_mcp_server_api_mcp_servers__name__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableMcpServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Mcp Server"
      }
    },
    "/api/mcp/servers/{name}/reconcile": {
      "post": {
        "operationId": "reconcile_mcp_server_api_mcp_servers__name__reconcile_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconcileMcpServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpApplyConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile Mcp Server"
      }
    },
    "/api/mcp/servers/{name}/set-harnesses": {
      "post": {
        "operationId": "set_mcp_server_harnesses_api_mcp_servers__name__set_harnesses_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMcpServerHarnessesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpSetHarnessesResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Mcp Server Harnesses"
      }
    },
    "/api/mcp/unmanaged/adopt": {
      "post": {
        "operationId": "adopt_mcp_server_api_mcp_unmanaged_adopt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptMcpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpApplyConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adopt Mcp Server"
      }
    },
...
  }
}